requirements.txt
opencv-python
easyocr
mss
pyyaml
numpy
